Initial Upload
Project was cleaned before upload and should work for different VS versions. No changes are needed, simply run the file and it should work out of the box.

The spec sheet is at the top of the Program.cs file. There are two versions of the solution in the file, the second is currently the active version. 
The first was an initial attempt to get a working version. The second contains improvements and should allow for easy creation of other products.